Changed Java wrapper datatypes to be compatible with JDK 1.4
SVN-Revision: 554

+++ src/gov/nih/nci/system/server/mgmt/SecurityEnabler.java
@@ -243,7 +243,7 @@ public class SecurityEnabler
 			catch (Exception ex)
 			{
 				log.error("No Security level found; setting a default security level");
-				securityLevel = SecurityConfiguration.getSecurityLevel();
+				securityLevel = SecurityConfiguration.getSecurityLevel().intValue();
 			}
 		}
 		return securityLevel;
